Migrated checkbox.go. I'll do more at once; I just want to see what effect this has.
